Keep track of polymorphic functions
Query TypeInference for the inferred type of the function of
interest and  check if that function's definition contains a free
variable. When it does, add it to the collection of functions
identified as polymorphic for later querying.

This introduces a new function on TipType and its inheritors to
identify if the type itself or any of its contents contain free
variables.

Comments make the kludges obvious.