Provide typing visitors with polymorphic functions
Type constraint visitors will be responsible for instantiating the
function typing for polymorphic functions. Take the set of found
polymorphic functions from the PolymorphicIdentifierVisitor and
give it to the TypConstraintCollectVisitor.

Change the unordered_set of polymorphic functions to be a std::set
since that is more prevalently used throughout the codebase, making
it easier to work with.

There are also a lot of whitespace changes here to conform with
what seems to be the prevalent style.