Label instances to separate constraint definitions
Append numeric identifiers to differentiate instantiations of
polymorphic functions. This only needs to apply to free variables.

Apply the parameters' type variables' constraints separately in
the constraint visitor to tie them to the separate free variables.

When encountering a polymorphic function, create the instantiation
and apply the constraints to the instantiation so they can be
solved independent of the original function definition.