Handle nested free variables
Free variables may not directly be parameters or return values.
They can end up nested in references or other cons types.

For example, the following program is problematic because the
parameter to f is a TipRef->TipAlpha and the return type is a
TipAlpha. Instantiation code did not work properly because it would
not instantiate the two free variables correctly and make them
refer to the same type variable.

f(a) {
  return *a;
}

main() {
  var x, y;
  x = alloc 2;
  y = f(x) + 1;

  return 0;
}

Add functions to retrieve the free variables in a type and replace
the free variables in a type. Replacement compares free variables
by name. When the names match, the function replaces the free
variable with the given replacement. Match on name only because
the free variables' nodes from the AST do not have the same pointer
even if they are referring to the same node.